test(deploy): share complete HTTP request drain
