+++ src/commands/mod.rs
@@ -14,6 +14,8 @@ pub(crate) mod run_state;
 pub(crate) mod self_test;
 pub(crate) mod setup;
 pub(crate) mod spel;
+#[cfg(test)]
+pub(crate) mod test_support;
 pub(crate) mod testnode;
 pub(crate) mod wallet;
 pub(crate) mod wallet_support;
