+++ src/commands/wallet_support.rs
@@ -527,10 +527,9 @@ fn one_line(text: &str) -> String {
 #[cfg(test)]
 mod tests {
     use std::fs;
-    use std::io::{Read, Write};
-    use std::net::{TcpListener, TcpStream};
+    use std::io::Write;
+    use std::net::TcpListener;
     use std::thread::JoinHandle;
-    use std::time::Duration;
 
     use tempfile::tempdir;
 
@@ -540,6 +539,7 @@ mod tests {
         read_default_wallet_address, resolve_wallet_address, set_wallet_home_env,
         wallet_state_path, write_default_wallet_address, WALLET_CONFIG_PRIMARY,
     };
+    use crate::commands::test_support::drain_http_request;
     use crate::constants::WALLET_HOME_ENV_VARS;
 
     const ACCOUNT_ID: &str = "6iArKUXxhUJqS7kCaPNhwMWt3ro71PDyBj7jwAyE2VQV";
@@ -563,45 +563,6 @@ mod tests {
         });
 
         (url, handle)
-    }
-
-    fn drain_http_request(stream: &mut TcpStream) {
-        let _ = stream.set_read_timeout(Some(Duration::from_secs(2)));
-        let mut request = Vec::new();
-        let mut buf = [0_u8; 1024];
-        loop {
-            match stream.read(&mut buf) {
-                Ok(0) => break,
-                Ok(n) => {
-                    request.extend_from_slice(&buf[..n]);
-                    if http_request_complete(&request) {
-                        break;
-                    }
-                }
-                Err(err)
-                    if matches!(
-                        err.kind(),
-                        std::io::ErrorKind::WouldBlock | std::io::ErrorKind::TimedOut
-                    ) =>
-                {
-                    break
-                }
-                Err(err) => panic!("read request: {err}"),
-            }
-        }
-    }
-
-    fn http_request_complete(request: &[u8]) -> bool {
-        let Some(header_end) = request.windows(4).position(|w| w == b"\r\n\r\n") else {
-            return false;
-        };
-        let headers = String::from_utf8_lossy(&request[..header_end]).to_ascii_lowercase();
-        let content_len = headers
-            .lines()
-            .find_map(|line| line.strip_prefix("content-length:"))
-            .and_then(|value| value.trim().parse::<usize>().ok())
-            .unwrap_or(0);
-        request.len() >= header_end + 4 + content_len
     }
 
     #[test]
